requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch
pydantic==2.5.2
jinja2==3.1.2
aiofiles==23.2.1